reversed patch - faszom (C++) .so-ba
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@3549 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ loader/wine/windef.h
@@ -151,7 +151,7 @@ typedef int             INT;
 typedef unsigned int    UINT;
 typedef unsigned short  WORD;
 typedef unsigned int    DWORD;
-typedef unsigned long   ULONG;
+typedef unsigned int    ULONG;
 typedef unsigned char   BYTE;
 typedef long            LONG;
 typedef short           SHORT;
